renamed shallowCopy.js to helper.js
